requirements.txt
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
coveralls==1.8.1
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.3
django-crispy-forms==1.7.2
django-import-export==1.2.0
djangorestframework==3.10.1
docopt==0.6.2
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
python-coveralls==2.9.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tablib==0.13.0
urllib3==1.25.3
xlrd==1.2.0
xlwt==1.3.0
